FCB-24: cod refactor after introduce the new desig
