+++ src/views/LearnSection.jsx
@@ -1,4 +1,4 @@
-import { Button } from '../../components/Button';
+import { Button } from '../components/Button';
 
 export const LearnSection = () => {
   return (
