--- src/views/userSections/actionsSection/actionsSection.style.scss
@@ -1,48 +1,0 @@
-@use '../../../constant/variables' as *;
-
-.actions {
-   width: 100vw;
-   padding: $g-medium;
-   display: flex;
-   flex-direction: column;
-   
-   border-radius: $g-medium;
-   background-color: $base-color;
-
-   &__buttons {
-      display: flex;
-      font-size: $f-normal ;
-      border-bottom: solid $g-extra-small $primary-color;
-      border-top: solid $g-extra-small $primary-color;
-   }
-}
-
-.action__button {
-   width: 50vw;
-}
-
-.add__form {
-   display: flex;
-   flex-direction: column;
-   align-items: center;
-
-   &__buttons {
-      // width: 230px;
-      padding-top: $g-large;
-      align-self: flex-end;
-
-      display: flex;
-      justify-content: space-around;
-
-   }
-}
-
-.learn__section {
-   padding: $g-extra-large;
-   
-   display: flex;
-   flex-direction: column;
-   
-   background-color: $base-color;
-   border-radius: $g-medium;
-}
