+++ src/views/wordsSection/FullCardView.jsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
-import expandArrow from '../../../assets/arrow-down-sign-navigate.png';
-import { Button } from '../../../components/Button';
+import expandArrow from '../../assets/arrow-down-sign-navigate.png';
+import { Button } from '../../components/Button';
 import { EditFullCardView } from './EditFullCardView';
 
 export const FullCardView = ({ cardPosition, cardData, id, onClose }) => {
