+++ src/views/wordsSection/WordCard.jsx
