+++ src/views/wordsSection/WordsSectionHeader.jsx
